Site Health: Remove the WordPress 5.2 reference from the email sent on fatal errors.
The version reference was initially added to reduce confusion and make it clear that this is new WordPress functionality, with the website itself sending an automated notification of a fatal error, most likely in a plugin or theme.

Almost four years and several major releases later, it no longer adds any helpful information and can be removed.

Follow-up to [44973], [45181], [45268].

Props NekoJonez, jb510, peterwilsoncc, kebbet, mukesh27, mrwweb, johnbillion, SergeyBiryukov.
Fixes #57327. See #54961.

git-svn-id: https://develop.svn.wordpress.org/trunk@55001 602fd350-edb4-49c9-b593-d223f7449a82